use op_word to store selected word

+++ jaylib.cpp
@@ -3,7 +3,7 @@
 
 void Jaylib::drawBand(uint8_t x, uint8_t y, const uint8_t * sprite, uint8_t cols) {
     uint16_t s;
-    int p = (y >> 3) * WIDTH + cols + x - 1;
+    uint16_t p = (y >> 3) * WIDTH + cols + x - 1;
 
     for(; cols --; p --) {
         s = pgm_read_byte(sprite + cols) << (y & 0x7);
